Assignment 9: Roberto Rodriguez
I have applied the Factory pattern to the CsvJsonConverter class.
Initially this class ran it's main method using a hardcoded string for a
csv data file to create a corresponding json file filled with data to be
used in the project.

Instead, a Converter interface was created (which was implemented by
CsvJsonConverter). This allows for a variety of converters of different
formats to be created (e.g. a CsvXmlConverter). Also, a ConverterFactory
class was created so that any code wanting to use a particular converter,
can instantiate a factory and then call it's getConverter method to get
the desired converter.